use standard python test file prefix
